requirements.txt
google-cloud-aiplatform>=1.70.0
pydantic>=2.0.0
python-dotenv
